+++ backend/tests/api/test_trade_order_detail_delete.py
@@ -9,7 +9,7 @@ from tests.factory.permission_factory import PermissionFactory
 
 
 @pytest.mark.django_db
-class TestOrderDeleteView:
+class TestTradeOrderDetailDelete:
     """
     注文削除APIのテスト
