+++ backend/tests/api/test_trade_order_detail_put.py
@@ -5,15 +5,16 @@ from apps.trade_flow.models.orders import Order
 from tests.factory.user_factory import UserFactory
 from tests.factory.permission_factory import PermissionFactory
 from tests.factory.customer_factory import CustomerFactory
+from tests.factory.order_factory import OrderFactory
 
 
 @pytest.mark.django_db
-class TestOrderCreateView:
+class TestTradeOrderDetailPut:
     """
-    注文情報を新規作成するビューのテスト
+    注文情報を更新するビューのテスト
 
-    url: /api/trade/orders/
-    method: POST
+    url: /api/trade/orders/<int:pk>/
+    method: PUT
     """
 
     @pytest.fixture
@@ -27,20 +28,24 @@ class TestOrderCreateView:
         return user
 
     @pytest.fixture
+    def order(self, customer):
+        return OrderFactory(customer=customer)
+
+    @pytest.fixture
     def customer(self):
         return CustomerFactory()
 
-    def test_create_order_success(self, client, authed_user, customer):
+    def test_update_order_success(self, client, authed_user, customer, order):
         """
-        正常系: 注文情報を新規作成する
+        正常系: 注文情報を更新する
 
         条件:
         - ユーザーが認証されている
         - 顧客情報が存在する
         
         結果:
-        - ステータスコード 201
-        - 注文情報が作成される
+        - ステータスコード 200
+        - 注文情報が更新される
         """
         client.force_authenticate(user=authed_user)
         payload = {
@@ -53,12 +58,12 @@ class TestOrderCreateView:
             "deadline": "2025-04-30",
             "note": "特記事項なし"
         }
-        response = client.post("/api/trade/orders/", data=payload, format="json")
-        assert response.status_code == status.HTTP_201_CREATED
+        response = client.put(f"/api/trade/orders/{order.id}/", data=payload, format="json")
+        assert response.status_code == status.HTTP_200_OK
         assert Order.objects.count() == 1
         assert response.data["order_number"] == "ORD-001"
 
-    def test_create_order_unauthorized(self, client, customer):
+    def test_update_order_unauthorized(self, client, customer, order):
         """
         異常系: 未認証
 
@@ -67,7 +72,7 @@ class TestOrderCreateView:
         
         結果:
         - ステータスコード 401
-        - 注文情報が作成されない
+        - 注文情報が更新されない
         """
         payload = {
             "customer": customer.id,
@@ -78,10 +83,10 @@ class TestOrderCreateView:
             "price": 3000,
             "deadline": "2025-04-20",
         }
-        response = client.post("/api/trade/orders/", data=payload, format="json")
+        response = client.put(f"/api/trade/orders/{order.id}/", data=payload, format="json")
         assert response.status_code == status.HTTP_401_UNAUTHORIZED
 
-    def test_create_order_forbidden(self, client, customer):
+    def test_update_order_forbidden(self, client, customer, order):
         """
         異常系: 編集権限がない
 
@@ -90,7 +95,7 @@ class TestOrderCreateView:
 
         結果:
         - ステータスコード 403
-        - 注文情報が作成されない
+        - 注文情報が更新されない
         """
         user = UserFactory()
         PermissionFactory(user=user, can_edit_order=False)
@@ -104,10 +109,10 @@ class TestOrderCreateView:
             "price": 1500,
             "deadline": "2025-04-15",
         }
-        response = client.post("/api/trade/orders/", data=payload, format="json")
+        response = client.put(f"/api/trade/orders/{order.id}/", data=payload, format="json")
         assert response.status_code == status.HTTP_403_FORBIDDEN
 
-    def test_create_order_invalid_data(self, client, authed_user, customer):
+    def test_update_order_invalid_data(self, client, authed_user, customer, order):
         """
         異常系: 不正なデータ
 
@@ -116,7 +121,7 @@ class TestOrderCreateView:
         
         結果:
         - ステータスコード 400
-        - 注文情報が作成されない
+        - 注文情報が更新されない
         """
         client.force_authenticate(user=authed_user)
         payload = {
@@ -128,5 +133,5 @@ class TestOrderCreateView:
             "price": "free",
             "deadline": "not-a-date",
         }
-        response = client.post("/api/trade/orders/", data=payload, format="json")
+        response = client.put(f"/api/trade/orders/{order.id}/", data=payload, format="json")
         assert response.status_code == status.HTTP_400_BAD_REQUEST
