+++ backend/tests/api/test_trade_order_post.py
@@ -5,16 +5,15 @@ from apps.trade_flow.models.orders import Order
 from tests.factory.user_factory import UserFactory
 from tests.factory.permission_factory import PermissionFactory
 from tests.factory.customer_factory import CustomerFactory
-from tests.factory.order_factory import OrderFactory
 
 
 @pytest.mark.django_db
-class TestOrderUpdateView:
+class TestTradeOrderPost:
     """
-    注文情報を更新するビューのテスト
+    注文情報を新規作成するビューのテスト
 
-    url: /api/trade/orders/<int:pk>/
-    method: PUT
+    url: /api/trade/orders/
+    method: POST
     """
 
     @pytest.fixture
@@ -28,24 +27,20 @@ class TestOrderUpdateView:
         return user
 
     @pytest.fixture
-    def order(self, customer):
-        return OrderFactory(customer=customer)
-
-    @pytest.fixture
     def customer(self):
         return CustomerFactory()
 
-    def test_update_order_success(self, client, authed_user, customer, order):
+    def test_create_order_success(self, client, authed_user, customer):
         """
-        正常系: 注文情報を更新する
+        正常系: 注文情報を新規作成する
 
         条件:
         - ユーザーが認証されている
         - 顧客情報が存在する
         
         結果:
-        - ステータスコード 200
-        - 注文情報が更新される
+        - ステータスコード 201
+        - 注文情報が作成される
         """
         client.force_authenticate(user=authed_user)
         payload = {
@@ -58,12 +53,12 @@ class TestOrderUpdateView:
             "deadline": "2025-04-30",
             "note": "特記事項なし"
         }
-        response = client.put(f"/api/trade/orders/{order.id}/", data=payload, format="json")
-        assert response.status_code == status.HTTP_200_OK
+        response = client.post("/api/trade/orders/", data=payload, format="json")
+        assert response.status_code == status.HTTP_201_CREATED
         assert Order.objects.count() == 1
         assert response.data["order_number"] == "ORD-001"
 
-    def test_update_order_unauthorized(self, client, customer, order):
+    def test_create_order_unauthorized(self, client, customer):
         """
         異常系: 未認証
 
@@ -72,7 +67,7 @@ class TestOrderUpdateView:
         
         結果:
         - ステータスコード 401
-        - 注文情報が更新されない
+        - 注文情報が作成されない
         """
         payload = {
             "customer": customer.id,
@@ -83,10 +78,10 @@ class TestOrderUpdateView:
             "price": 3000,
             "deadline": "2025-04-20",
         }
-        response = client.put(f"/api/trade/orders/{order.id}/", data=payload, format="json")
+        response = client.post("/api/trade/orders/", data=payload, format="json")
         assert response.status_code == status.HTTP_401_UNAUTHORIZED
 
-    def test_update_order_forbidden(self, client, customer, order):
+    def test_create_order_forbidden(self, client, customer):
         """
         異常系: 編集権限がない
 
@@ -95,7 +90,7 @@ class TestOrderUpdateView:
 
         結果:
         - ステータスコード 403
-        - 注文情報が更新されない
+        - 注文情報が作成されない
         """
         user = UserFactory()
         PermissionFactory(user=user, can_edit_order=False)
@@ -109,10 +104,10 @@ class TestOrderUpdateView:
             "price": 1500,
             "deadline": "2025-04-15",
         }
-        response = client.put(f"/api/trade/orders/{order.id}/", data=payload, format="json")
+        response = client.post("/api/trade/orders/", data=payload, format="json")
         assert response.status_code == status.HTTP_403_FORBIDDEN
 
-    def test_update_order_invalid_data(self, client, authed_user, customer, order):
+    def test_create_order_invalid_data(self, client, authed_user, customer):
         """
         異常系: 不正なデータ
 
@@ -121,7 +116,7 @@ class TestOrderUpdateView:
         
         結果:
         - ステータスコード 400
-        - 注文情報が更新されない
+        - 注文情報が作成されない
         """
         client.force_authenticate(user=authed_user)
         payload = {
@@ -133,5 +128,5 @@ class TestOrderUpdateView:
             "price": "free",
             "deadline": "not-a-date",
         }
-        response = client.put(f"/api/trade/orders/{order.id}/", data=payload, format="json")
+        response = client.post("/api/trade/orders/", data=payload, format="json")
         assert response.status_code == status.HTTP_400_BAD_REQUEST
